Remove module 'utils/schedule'
The functionality of the module 'utils/schedule' was moved to modules 'handlers', 'utils/beautify' and 'utils/api'. It also simplifies testing.

Function 'dayName' of module 'utils/date' now accepts index of the day (from 0 to 6) and returns name of the week's day. Note: day with index '0' is Monday, not Sunday.